+++ sentry/src/main/java/io/sentry/Hub.java
@@ -600,7 +600,7 @@ public final class Hub implements IHub {
   @Override
   public @NotNull SentryId captureTransaction(
       final @NotNull SentryTransaction transaction,
-      final @Nullable TraceState traceState,
+      final @Nullable TraceContext traceContext,
       final @Nullable Hint hint,
       final @Nullable ProfilingTraceData profilingTraceData) {
     Objects.requireNonNull(transaction, "transaction is required");
@@ -638,7 +638,7 @@ public final class Hub implements IHub {
             sentryId =
                 item.getClient()
                     .captureTransaction(
-                        transaction, traceState, item.getScope(), hint, profilingTraceData);
+                        transaction, traceContext, item.getScope(), hint, profilingTraceData);
           } catch (Throwable e) {
             options
                 .getLogger()