+++ sentry/src/main/java/io/sentry/NoOpHub.java
@@ -136,7 +136,7 @@ public final class NoOpHub implements IHub {
   @Override
   public @NotNull SentryId captureTransaction(
       final @NotNull SentryTransaction transaction,
-      final @Nullable TraceState traceState,
+      final @Nullable TraceContext traceContext,
       final @Nullable Hint hint,
       final @Nullable ProfilingTraceData profilingTraceData) {
     return SentryId.EMPTY_ID;