+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -1189,7 +1189,7 @@ class HubTest {
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op"), sut)
         sentryTracer.finish()
-        sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceState?)
+        sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceContext?)
         verify(mockClient, never()).captureTransaction(any(), any(), any())
         verify(mockClient, never()).captureTransaction(any(), any(), any(), anyOrNull(), anyOrNull())
     }
@@ -1206,8 +1206,8 @@ class HubTest {
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
         sentryTracer.finish()
-        val traceState = sentryTracer.traceState()
-        verify(mockClient).captureTransaction(any(), eq(traceState), any(), eq(null), anyOrNull())
+        val traceContext = sentryTracer.traceContext()
+        verify(mockClient).captureTransaction(any(), eq(traceContext), any(), eq(null), anyOrNull())
     }
 
     @Test
@@ -1273,7 +1273,7 @@ class HubTest {
         sut.bindClient(mockClient)
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", true), sut)
-        sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceState?)
+        sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceContext?)
         verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null), anyOrNull())
     }
 
@@ -1289,8 +1289,8 @@ class HubTest {
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", false), sut)
         sentryTracer.finish()
-        val traceState = sentryTracer.traceState()
-        verify(mockClient, never()).captureTransaction(any(), eq(traceState), any(), eq(null), anyOrNull())
+        val traceContext = sentryTracer.traceContext()
+        verify(mockClient, never()).captureTransaction(any(), eq(traceContext), any(), eq(null), anyOrNull())
     }
 
     @Test